requirements.txt
# Existing dependencies
SpeechRecognition==3.10.0
RPi.GPIO==0.7.1

# Azure AI services
openai==1.35.0
azure-cognitiveservices-speech==1.30.0

# Text-to-speech
pyttsx3==2.90

# Computer Vision
opencv-python==4.8.1.78
ultralytics==8.0.196
picamera2==0.3.12
numpy==1.24.3
pillow==10.0.0

# Additional utilities
python-dotenv==1.0.0
requests==2.31.0